Use -cp so that we only need to generate and upload a single JAR.
