+++ load-test-framework/pom.xml
@@ -165,75 +165,6 @@
             <finalName>driver</finalName>
           </configuration>
           </execution>
-          <execution>
-            <id>gcloud_publisher</id>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-            <configuration>
-              <createDependencyReducedPom>false</createDependencyReducedPom>
-              <transformers>
-                <!-- add Main-Class to manifest file -->
-                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <mainClass>com.google.pubsub.clients.gcloud.CPSPublishingTask</mainClass>
-                </transformer>
-              </transformers>
-              <finalName>gcloud_publisher</finalName>
-            </configuration>
-          </execution>
-          <execution>
-            <id>gcloud_subscriber</id>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-            <configuration>
-              <createDependencyReducedPom>false</createDependencyReducedPom>
-              <transformers>
-                <!-- add Main-Class to manifest file -->
-                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <mainClass>com.google.pubsub.clients.gcloud.CPSSubscriberTask</mainClass>
-                </transformer>
-              </transformers>
-              <finalName>gcloud_subscriber</finalName>
-            </configuration>
-          </execution>
-          <execution>
-            <id>kafka_publisher</id>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-            <configuration>
-              <createDependencyReducedPom>false</createDependencyReducedPom>
-              <transformers>
-                <!-- add Main-Class to manifest file -->
-                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <mainClass>com.google.pubsub.clients.kafka.KafkaPublishingTask</mainClass>
-                </transformer>
-              </transformers>
-              <finalName>kafka_publisher</finalName>
-            </configuration>
-          </execution>
-          <execution>
-            <id>kafka_subscriber</id>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-            <configuration>
-              <createDependencyReducedPom>false</createDependencyReducedPom>
-              <transformers>
-                <!-- add Main-Class to manifest file -->
-                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <mainClass>com.google.pubsub.clients.kafka.KafkaSubscriberTask</mainClass>
-                </transformer>
-              </transformers>
-              <finalName>kafka_subscriber</finalName>
-            </configuration>
-          </execution>
-
         </executions>
        </plugin>
       <plugin>
