--- load-test-framework/src/main/resources/consumer.properties
@@ -1,5 +1,0 @@
-key.deserializer=org.apache.kafka.common.serialization.StringDeserializer
-value.deserializer=org.apache.kafka.common.serialization.StringDeserializer
-group.id=SUBSCRIBER_ID
-enable.auto.commit=true
-session.timeout.ms=30000
