+++ load-test-framework/src/main/resources/gce/kafka-publisher_startup_script.sh
@@ -4,7 +4,7 @@ readonly TMP="$(mktemp -d)"
 [[ "${TMP}" != "" ]] || error mktemp failed
 
 # Download the loadtest binary to this machine.
-/usr/bin/gsutil cp "gs://cloud-pubsub-loadtest/kafka_publisher.jar" "${TMP}"
+/usr/bin/gsutil cp "gs://cloud-pubsub-loadtest/driver.jar" "${TMP}"
 
 # Run the loadtest binary
-java -Xmx5000M -jar ${TMP}/kafka_publisher.jar
+java -Xmx5000M -cp ${TMP}/driver.jar com.google.pubsub.clients.kafka.KafkaPublisherTask
