--- load-test-framework/src/main/resources/producer.properties
@@ -1,4 +1,0 @@
-max.block.ms = 30000
-key.serializer=org.apache.kafka.common.serialization.StringSerializer
-value.serializer=org.apache.kafka.common.serialization.StringSerializer
-acks=all
